pyproject.toml
[project]
name = "bananarama"
version = "0.1.0"
description = "Generate presentation images from YAML configs using AI image models"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pillow>=10.0",
    "click>=8.0",
    "rich>=13.0",
    "google-genai>=1.0",
]

[project.optional-dependencies]
openai = ["openai>=1.50"]
flux = ["together>=1.0"]
all = ["openai>=1.50", "together>=1.0"]

[project.scripts]
bananarama = "bananarama.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bananarama"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["bananarama"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["google.*", "PIL.*", "together.*", "openai.*", "httpx.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: real API calls (require API keys, deselected by default)",
]
addopts = "-m 'not integration'"
